Make list items shared component
