+++ cascade_ui/web/src/views/ModelLineView.vue
@@ -3,89 +3,23 @@ import NavBar from "../components/NavBar.vue";
 import GetRepo from "@/components/GetRepo";
 import GetLine from "@/components/GetLine";
 import GetWorkspace from "@/components/GetWorkspace";
+import ListItems from "@/components/ListItems.vue";
 import { ref, onMounted, computed } from "vue";
 import { Repo as RepoClass } from "@/models/Repo";
 import {ModelLine} from "@/models/ModelLine";
 import type {Repo} from "@/models/Repo";
 import type {Workspace} from "@/models/Workspace";
 import { Workspace as WorkspaceClass } from "@/models/Workspace";
-import { useRouter, useRoute } from 'vue-router'
+import { useRoute } from 'vue-router'
 
 const route = useRoute()
-const router = useRouter()
+
 const repoName = computed(() => route.params.repoName as string)
 const lineName = computed(() => route.params.lineName as string)
 
 const workspace = ref<Workspace | null>(null);
 const repo = ref<Repo | null>(null);
 const line = ref<ModelLine | null>(null);
-
-const selectedFields = ref<string[]>([]);
-const requestClicked = ref(false);
-
-const defaultFields = ['name', 'slug', 'tags', 'created_at', 'saved_at'];
-
-async function fetchLineItems() {
-  if (!repoName.value || !lineName.value) return;
-  const fields = Array.from(new Set(selectedFields.value));
-  const response = await fetch("/v1/line_item_table", {
-    method: "POST",
-    headers: { "Content-Type": "application/json" },
-    body: JSON.stringify({
-      line_path: { repo: repoName.value, line: lineName.value },
-      item_fields: fields,
-    }),
-  });
-  if (response.ok) {
-    const items = await response.json();
-    if (line.value) {
-      const filledItems = items.map((item: Record<string, any>, idx: number) => {
-        let existing: Record<string, any> | undefined = undefined;
-        if (line.value && Array.isArray(line.value.items)) {
-          existing = line.value.items[idx] as Record<string, any>;
-        }
-        for (const field of defaultFields) {
-          if (!(field in item)) {
-            if (existing && field in existing) {
-              item[field] = existing[field];
-            } else {
-              item[field] = "";
-            }
-          }
-        }
-        return item;
-      });
-      line.value.items = filledItems;
-    }
-  }
-}
-
-const fieldsOptions = computed(() => {
-  return line.value?.item_fields.map(field => field.toString()) || [];
-});
-
-const dynamicModelHeaders = computed(() => {
-  const baseHeaders = [
-    { title: 'name', value: 'name' },
-    { title: 'slug', value: 'slug' },
-    { title: 'tags', value: 'tags' },
-    { title: 'created', value: 'created_at' },
-    { title: 'saved', value: 'saved_at' },
-  ];
-  const selected = selectedFields.value
-    .filter(f => !baseHeaders.some(h => h.value === f))
-    .map(f => ({ title: f, value: f }));
-  return baseHeaders.concat(selected);
-});
-
-function applyFields() {
-  requestClicked.value = true;
-  fetchLineItems();
-}
-
-function openModel(repoName: string, lineName: string, modelNumString: string) {
-  router.push({ name: "model", params: { repoName, lineName, modelNumString } });
-}
 
 onMounted(async () => {
   const wsObj = await GetWorkspace();
@@ -128,36 +62,7 @@ const breadcrumbs = computed(() => {
     <div>
       <div class="content">
         <v-breadcrumbs :items="breadcrumbs"></v-breadcrumbs>
-        <div v-if="line && line.items">
-          <div class="mb-4" style="display: flex; align-items: center; gap: 16px;">
-            <v-select
-              v-model="selectedFields"
-              :items="fieldsOptions"
-              label="Select columns"
-              multiple
-              chips
-              item-title="."
-              item-value="."
-              :menu-props="{ closeOnContentClick: false }"
-              persistent-hint
-              hint="Choose which additional columns to display"
-              prepend-icon="mdi-table-column"
-            >
-              <template #item="{ props }">
-                <v-list-item v-bind="props">
-                </v-list-item>
-              </template>
-            </v-select>
-            <v-btn color="primary" @click="applyFields">Request</v-btn>
-          </div>
-          <v-data-table :headers="dynamicModelHeaders" :items="line.items" class="mt-4">
-            <template #item.name="{ item }">
-              <v-btn variant="text" style="font-family: Roboto,serif; font-size: 14px; color: #DEB841;" @click="openModel(repoName, lineName, item.name)">
-                {{ item.name }}
-              </v-btn>
-            </template>
-          </v-data-table>
-        </div>
+        <ListItems v-if="line" :line="line"/>
       </div>
     </div>
   </body>
